accounts: move mute stuff to own module
