+++ crates/notedeck/src/account/mod.rs
@@ -1,2 +1,3 @@
 pub mod accounts;
+pub mod mute;
 pub mod relay;
